refactor: add auth router
