+++ src/routes/indext.ts
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+
+import authRouter from './authRouter';
+
+const router = Router();
+
+router.use('/auth', authRouter);
+
+export default router;
